Add validation

+++ controllers/cards.js
@@ -1,4 +1,7 @@
+const mongoose = require('mongoose');
 const Card = require('../models/card');
+
+const { ObjectId } = mongoose.Types;
 
 module.exports.getCards = (req, res) => {
   Card.find({})
@@ -10,20 +13,29 @@ module.exports.getCards = (req, res) => {
 module.exports.createCard = (req, res) => {
   const owner = req.user;
   const { name, link } = req.body;
-
-  Card.create({ name, link, owner })
-    .then((card) => res.send({ data: card }))
-    .catch((err) => res.status(500).send({ message: err.message }));
+  const newCard = { name, link, owner };
+  const error = new Card(newCard).validateSync();
+  if (error) {
+    res.status(400).send({ message: `Ошибка валидации ${error}` });
+  } else {
+    Card.create(newCard)
+      .then((card) => res.send({ data: card }))
+      .catch((err) => res.status(500).send({ message: err.message }));
+  }
 };
 
 module.exports.deleteCard = (req, res) => {
-  Card.findOneAndDelete({ _id: req.params.cardId, owner: req.user._id })
-    .then((card) => {
-      if (card) {
-        res.send({ data: card });
-      } else {
-        res.status(404).send({ message: 'Карточка не найдена' });
-      }
-    })
-    .catch((err) => res.status(500).send({ message: err.message }));
+  if (!ObjectId.isValid(req.params.cardId)) {
+    res.status(400).send({ message: 'Ошибка валидации' });
+  } else {
+    Card.findOneAndDelete({ _id: req.params.cardId, owner: req.user._id })
+      .then((card) => {
+        if (card) {
+          res.send({ data: card });
+        } else {
+          res.status(404).send({ message: 'Карточка не найдена' });
+        }
+      })
+      .catch((err) => res.status(500).send({ message: err.message }));
+  }
 };
